webkit/data/layout_tests/chrome/fast/dom/Window/customized-property-survives-gc.html

<html>
<script>
if (window.layoutTestController)
  layoutTestController.dumpAsText();

function CollectGarbage() {
  if (window.gc) {
    // Chrome code
    window.gc();
  } else {
    // Safari hack
    for (var i = 0; i < 50000; i++)
      new Object();
  }
}


function check(name) {
  if (!window[name]) return;

  window[name].myProp = 10;
  CollectGarbage();
  var r = document.getElementById("result");
  if (window[name].myProp) {
    r.innerHTML += name + ".myProp survived GC.<br>";
  } else {
    r.innerHTML += name + ".myProp did not survive GC.<br>";
  }
}

function runTest() {
  check("screen");
  check("history");
  check("locationbar");
  check("menubar");
  check("personalbar");
  check("scrollbars");
  check("statusbar");
  check("toolbar");
  check("location");
  check("navigator");
}
</script>

<body onload="runTest()">
This tests that customized properties on window.location and window.navigator
won't get lost after a GC.
<div id="result">TEST running.<br></div>
</body>
</html>
