webkit/data/layout_tests/chrome/fast/dom/Window/customized-property-survives-gc-expected.txt

This tests that customized properties on window.location and window.navigator won't get lost after a GC.
TEST running.
screen.myProp did not survive GC.
history.myProp survived GC.
locationbar.myProp did not survive GC.
menubar.myProp did not survive GC.
personalbar.myProp did not survive GC.
scrollbars.myProp did not survive GC.
statusbar.myProp did not survive GC.
toolbar.myProp did not survive GC.
location.myProp survived GC.
navigator.myProp survived GC.
